feat: change return type data class file location
