+++ src/main/kotlin/org/appjam/smashing/domain/review/dto/response/ReviewRatingCount.kt
@@ -1,0 +1,8 @@
+package org.appjam.smashing.domain.review.dto.response
+
+import org.appjam.smashing.domain.review.enums.ReviewRating
+
+data class ReviewRatingCount(
+    val reviewRating: ReviewRating?,
+    val counts: Long?
+)
